auth/krb5: move kerberos5 wrapper to rust root
Make it available outside of just the SMB parser.
