+++ rust/src/smb/mod.rs
@@ -15,8 +15,6 @@
  * 02110-1301, USA.
  */
 
-extern crate kerberos_parser;
-
 pub mod smb_records;
 pub mod smb1_records;
 pub mod smb2_records;
